+++ code/observers_and_api/src/Hdf5Observer.cpp
@@ -81,9 +81,9 @@ std::function<void()> Hdf5Observer::get_initializer(const SurfaceElevationGrid& 
 {
     return [this,waveElevationGrid, addressing]()
            {
-               const size_t nx = waveElevationGrid.x.size();
-               const size_t ny = waveElevationGrid.y.size();
-               wave_serializer = TR1(shared_ptr)<SimHdf5WaveObserver>(new SimHdf5WaveObserver(h5File, this->basename+"/waves",nx,ny));
+               const long nx = (long)waveElevationGrid.x.size();
+               const long ny = (long)waveElevationGrid.y.size();
+               wave_serializer = TR1(shared_ptr)<SimHdf5WaveObserver>(new SimHdf5WaveObserver(h5File, this->basename+"/waves",(size_t)nx,(size_t)ny));
            };
 }
